requirements.txt
# Core RAG dependencies
google-generativeai==0.8.3
pymongo==4.8.0
sentence-transformers==3.0.1
transformers>=4.30.0
torch>=2.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.1

# Evaluation frameworks
ragas>=0.2.0
deepeval>=1.1.5

# Data processing
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.5

# Document processing
markdown==3.7
beautifulsoup4==4.12.3

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.1
rich==13.8.1
click==8.1.7

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0
isort==5.13.2